backend/pyproject.toml
[project]
name = "veterinary-medical-records"
requires-python = ">=3.13"
version = "0.0.1"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.20",
    "pypdf>=5.1.0",
    "pytesseract>=0.3.13",
    "pillow>=10.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.36",
    "psycopg[binary]==3.2.3",
    "sqlmodel>=0.0.16",
    "alembic>=1.13.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.24.0,<0.25.0",
    "ruff>=0.14.9",
    "mypy>=1.13.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::PendingDeprecationWarning:starlette.formparsers",
]
